# docs/swagger.yaml
basePath: /
definitions:
...
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: integer
      item:
        type: string
      nickname:
        type: string
      price:
        type: number
      quantity:
        type: integer
      size:
        type: string
      status:
        type: string
      total_price:
        type: number
      user_id:
        type: integer
    type: object
  models.Product:
    properties:
      description:
        type: string
      id:
        type: integer
      image_url:
        description: Novo campo para armazenar o caminho da imagem
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: suporte@seusite.com
    name: Suporte da API
    url: http://www.seusite.com/support
  description: API para gerenciamento de produtos, pedidos e usuários.
  title: API de Vendas
  version: "1.0"
paths:
...
  /orders:
    post:
      consumes:
      - application/json
      description: Cria um novo pedido com os dados fornecidos
      parameters:
      - description: Dados do pedido
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao criar pedido
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um novo pedido
      tags:
      - Orders
  /orders/{order_id}:
    get:
      description: Retorna os detalhes de um pedido específico pelo ID
      parameters:
      - description: ID do Pedido
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao buscar detalhes do pedido
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtém detalhes de um pedido
      tags:
      - Orders
  /orders/{order_id}/cancel:
    post:
      description: Cancela um pedido específico pelo ID
      parameters:
      - description: ID do Pedido
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Pedido cancelado com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao cancelar pedido
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cancela um pedido
      tags:
      - Orders
  /orders/{order_id}/track:
    get:
      description: Obtém o status atual de um pedido específico pelo ID
      parameters:
      - description: ID do Pedido
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao rastrear pedido
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rastreia um pedido
      tags:
      - Orders
  /payments:
    post:
      consumes:
      - application/json
      description: Inicia o processo de pagamento via Stripe Checkout e retorna a
        URL do Checkout
      parameters:
      - description: Dados do pedido
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "200":
          description: checkout_url
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao criar pagamento
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um pagamento
      tags:
      - Pagamentos
  /payments/webhook:
    post:
      consumes:
      - application/json
      description: Recebe notificações do Stripe sobre o status do pagamento
      produces:
      - application/json
      responses:
        "200":
          description: Webhook recebido
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Falha ao processar webhook
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Webhook de pagamento
      tags:
      - Pagamentos
  /products:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de produtos disponíveis no catálogo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Erro ao listar produtos
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista todos os produtos
      tags:
      - Produtos
    post:
      consumes:
      - multipart/form-data
      description: Cria um novo produto enviando dados e uma imagem
      parameters:
      - description: Nome do produto
        in: formData
        name: name
        required: true
        type: string
      - description: Descrição do produto
        in: formData
        name: description
        required: true
        type: string
      - description: Preço do produto
        in: formData
        name: price
        required: true
        type: number
      - description: Imagem do produto
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Produto criado com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Erro ao receber imagem
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao criar produto
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um produto
      tags:
      - Produtos
  /products/{product_id}:
    delete:
      consumes:
      - application/json
      description: Deleta um produto pelo ID
      parameters:
      - description: ID do produto
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Produto removido com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao remover produto
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove um produto
      tags:
      - Produtos
    get:
      consumes:
      - application/json
      description: Retorna um produto pelo seu ID
      parameters:
      - description: ID do produto
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "500":
          description: Erro ao recuperar produto
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtém um produto
      tags:
      - Produtos
    put:
      consumes:
      - multipart/form-data
      description: Atualiza as informações de um produto pelo ID
      parameters:
      - description: ID do produto
        in: path
        name: product_id
        required: true
        type: integer
      - description: Nome do produto
        in: formData
        name: name
        type: string
      - description: Descrição do produto
        in: formData
        name: description
        type: string
      - description: Preço do produto
        in: formData
        name: price
        type: number
      - description: Nova imagem do produto
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Produto atualizado com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao atualizar produto
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza um produto
      tags:
      - Produtos
...
  /users/{user_id}/orders:
    get:
      description: Retorna uma lista de pedidos de um usuário específico pelo ID
      parameters:
      - description: ID do Usuário
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "400":
          description: ID do usuário inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao buscar pedidos
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista pedidos do usuário
      tags:
      - Orders
swagger: "2.0"